Format the CDC files in place with xmllint
With the idea of future updates being repeatable, use the standard
utility "xmllint" to format the xml, since Eclipse would not be used
by all parties.

Also converted to use the line endings exactly as they download from
the CDC website

Simple script added to the directory, for a repeatable update process.